Move the store to lib
