+++ src/store/actions/PersonalDetailsActions.js
@@ -1,5 +1,5 @@
 import _ from 'underscore';
-import * as Store from '../Store';
+import * as Store from '../../lib/Store';
 import {request} from '../../lib/Network';
 import STOREKEYS from '../STOREKEYS';
 import md5 from '../../lib/md5';
